Profile, Edit Profile, Forgot Password, app_theme
the theme folder is required for the screens i made, the folder should be placed in lib folder (AET-Main/lib/theme). Login screen has been edited to navigate to forgot password screen and changed to use app_theme (does not need to be added if the original login screen can navigate to the forgot password screen easily later.